Refresh timeNow even before the first tick
Without this fix, when I run code similar to the one in #181 (i.e. displaying the amount of time that has passed since the button was clicked, which is _always_ in the past of course), then it displays "within a few seconds" at first, and only after a second (after the first tick fires) it updates to the correct "a few seconds ago". That's because right after the button click, `date` has already been moved forward, but `timeNow` has not. The change of `date` triggers the `useEffect`, but that doesn't trigger a `timeNow` update until the first tick fired. Therefore, add a `timeNow` update immediately as well.